requirements.txt
Flask
PyMuPDF
google-generativeai
